[update] #44 revivalをrestoreに変更、ディレクトリを新設
